+++ vowpal_platypus/internal.py
